feat: hide context menu when onRenderContextMenu return false
