+++ src/lib/components/modes/tablemode/TableMode.svelte
@@ -95,10 +95,13 @@
     getFocusPath,
     isEditingSelection,
     isJSONSelection,
+    isKeySelection,
+    isMultiSelection,
     isValueSelection,
     pathInSelection,
     pathStartsWith,
-    removeEditModeFromSelection
+    removeEditModeFromSelection,
+    singleItemSelected
   } from '$lib/logic/selection.js'
   import { createHistory } from '$lib/logic/history.js'
   import ColumnHeader from './ColumnHeader.svelte'
@@ -108,7 +111,18 @@
   import { getContext, onDestroy, onMount, tick } from 'svelte'
   import { jsonrepair } from 'jsonrepair'
   import Message from '../../controls/Message.svelte'
-  import { faCheck, faCode, faWrench } from '@fortawesome/free-solid-svg-icons'
+  import {
+    faCheck,
+    faClone,
+    faCode,
+    faCopy,
+    faCut,
+    faPaste,
+    faPen,
+    faPlus,
+    faTrashCan,
+    faWrench
+  } from '@fortawesome/free-solid-svg-icons'
   import { measure } from '$lib/utils/timeUtils.js'
   import memoizeOne from 'memoize-one'
   import { validateJSON } from '$lib/logic/validation.js'
@@ -142,6 +156,8 @@
   import JSONPreview from '../../controls/JSONPreview.svelte'
   import RefreshColumnHeader from './RefreshColumnHeader.svelte'
   import type { Context } from 'svelte-simple-modal'
+  import type { ContextMenuItem } from '$lib/types'
+  import { faCheckSquare, faSquare } from '@fortawesome/free-regular-svg-icons'
 
   const debug = createDebug('jsoneditor:TableMode')
   const { open } = getContext<Context>('simple-modal')
@@ -183,6 +199,29 @@
     escapeControlCharacters,
     escapeUnicodeCharacters
   })
+
+  $: selection = documentState.selection
+
+  $: hasJson = json !== undefined
+  $: hasSelection = !!selection
+  $: focusValue = json !== undefined && selection ? getIn(json, getFocusPath(selection)) : undefined
+
+  $: hasSelectionContents =
+    hasJson &&
+    (isMultiSelection(selection) || isKeySelection(selection) || isValueSelection(selection))
+
+  $: canEditValue = hasJson && selection != null && singleItemSelected(selection)
+  $: canEnforceString = canEditValue && !isObjectOrArray(focusValue)
+
+  $: enforceString =
+    selection != null && focusValue !== undefined
+      ? getEnforceString(
+          focusValue,
+          documentState.enforceStringMap,
+          compileJSONPointer(getFocusPath(selection)),
+          parser
+        )
+      : false
 
   let refJsonEditor: HTMLDivElement
   let refContents: HTMLDivElement | undefined
@@ -921,25 +960,179 @@
     offsetLeft,
     showTip
   }: AbsolutePopupOptions) {
+    let defaultItems: ContextMenuItem[] = [
+      { type: 'separator' },
+      {
+        type: 'row',
+        items: [
+          {
+            type: 'column',
+            items: [
+              { type: 'label', text: 'Table cell:' },
+              {
+                type: 'dropdown-button',
+                main: {
+                  type: 'button',
+                  onClick: () => handleEditValue(),
+                  icon: faPen,
+                  text: 'Edit',
+                  title: 'Edit the value (Double-click on the value)',
+                  disabled: !canEditValue
+                },
+                width: '11em',
+                items: [
+                  {
+                    type: 'button',
+                    icon: faPen,
+                    text: 'Edit',
+                    title: 'Edit the value (Double-click on the value)',
+                    onClick: () => handleEditValue(),
+                    disabled: !canEditValue
+                  },
+                  {
+                    type: 'button',
+                    icon: enforceString ? faCheckSquare : faSquare,
+                    text: 'Enforce string',
+                    title: 'Enforce keeping the value as string when it contains a numeric value',
+                    onClick: () => handleToggleEnforceString(),
+                    disabled: !canEnforceString
+                  }
+                ]
+              },
+              {
+                type: 'dropdown-button',
+                main: {
+                  type: 'button',
+                  onClick: () => handleCut(true),
+                  icon: faCut,
+                  text: 'Cut',
+                  title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
+                  disabled: !hasSelectionContents
+                },
+                width: '10em',
+                items: [
+                  {
+                    type: 'button',
+                    icon: faCut,
+                    text: 'Cut formatted',
+                    title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
+                    onClick: () => handleCut(true),
+                    disabled: !hasSelectionContents
+                  },
+                  {
+                    type: 'button',
+                    icon: faCut,
+                    text: 'Cut compacted',
+                    title: 'Cut selected contents, without indentation (Ctrl+Shift+X)',
+                    onClick: () => handleCut(false),
+                    disabled: !hasSelectionContents
+                  }
+                ]
+              },
+              {
+                type: 'dropdown-button',
+                main: {
+                  type: 'button',
+                  onClick: () => handleCopy(true),
+                  icon: faCopy,
+                  text: 'Copy',
+                  title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
+                  disabled: !hasSelectionContents
+                },
+                width: '12em',
+                items: [
+                  {
+                    type: 'button',
+                    icon: faCopy,
+                    text: 'Copy formatted',
+                    title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
+                    onClick: () => handleCopy(false),
+                    disabled: !hasSelectionContents
+                  },
+                  {
+                    type: 'button',
+                    icon: faCopy,
+                    text: 'Copy compacted',
+                    title: 'Copy selected contents, without indentation (Ctrl+Shift+C)',
+                    onClick: () => handleCopy(false),
+                    disabled: !hasSelectionContents
+                  }
+                ]
+              },
+              {
+                type: 'button',
+                onClick: () => handlePasteFromMenu(),
+                icon: faPaste,
+                text: 'Paste',
+                title: 'Paste clipboard contents (Ctrl+V)',
+                disabled: !hasSelection
+              },
+              {
+                type: 'button',
+                onClick: () => handleRemove(),
+                icon: faTrashCan,
+                text: 'Remove',
+                title: 'Remove selected contents (Delete)',
+                disabled: !hasSelectionContents
+              }
+            ]
+          },
+          {
+            type: 'column',
+            items: [
+              { type: 'label', text: 'Table row:' },
+              {
+                type: 'button',
+                onClick: () => handleEditRow(),
+                icon: faPen,
+                text: 'Edit row',
+                title: 'Edit the current row',
+                disabled: !hasSelectionContents
+              },
+              {
+                type: 'button',
+                onClick: () => handleDuplicateRow(),
+                icon: faClone,
+                text: 'Duplicate row',
+                title: 'Duplicate the current row',
+                disabled: !hasSelection
+              },
+              {
+                type: 'button',
+                onClick: () => handleInsertBeforeRow(),
+                icon: faPlus,
+                text: 'Insert before',
+                title: 'Insert a row before the current row',
+                disabled: !hasSelection
+              },
+              {
+                type: 'button',
+                onClick: () => handleInsertAfterRow(),
+                icon: faPlus,
+                text: 'Insert after',
+                title: 'Insert a row after the current row',
+                disabled: !hasSelection
+              },
+              {
+                type: 'button',
+                onClick: () => handleRemoveRow(),
+                icon: faTrashCan,
+                text: 'Remove row',
+                title: 'Remove current row',
+                disabled: !hasSelection
+              }
+            ]
+          }
+        ]
+      }
+    ]
+
+    const items = onRenderContextMenu(defaultItems)
+    if (items === false) return
+
     const props = {
-      json,
-      documentState: documentState,
-      parser,
       showTip,
-
-      onEditValue: handleEditValue,
-      onEditRow: handleEditRow,
-      onToggleEnforceString: handleToggleEnforceString,
-      onCut: handleCut,
-      onCopy: handleCopy,
-      onPaste: handlePasteFromMenu,
-      onRemove: handleRemove,
-      onDuplicateRow: handleDuplicateRow,
-      onInsertBeforeRow: handleInsertBeforeRow,
-      onInsertAfterRow: handleInsertAfterRow,
-      onRemoveRow: handleRemoveRow,
-
-      onRenderContextMenu,
+      items,
       onCloseContextMenu: function () {
         closeAbsolutePopup(popupId)
         focus()
@@ -965,7 +1158,7 @@
   }
 
   function handleContextMenu(event: Event) {
-    if (readOnly || isEditingSelection(documentState.selection)) {
+    if (isEditingSelection(documentState.selection)) {
       return
     }
 
@@ -1014,10 +1207,6 @@
   }
 
   function handleContextMenuFromTableMenu(event: MouseEvent) {
-    if (readOnly) {
-      return
-    }
-
     openContextMenu({
       anchor: findParentWithNodeName(event.target as HTMLElement, 'BUTTON'),
       offsetTop: 0,
