+++ src/lib/types.ts
@@ -314,8 +314,8 @@ export type RenderContextMenuContext = RenderMenuContext & { selection: JSONEdit
 export type OnRenderContextMenu = (
   items: ContextMenuItem[],
   context: RenderContextMenuContext
-) => ContextMenuItem[] | undefined
-export type OnRenderContextMenuInternal = (items: ContextMenuItem[]) => ContextMenuItem[]
+) => ContextMenuItem[] | false | undefined
+export type OnRenderContextMenuInternal = (items: ContextMenuItem[]) => ContextMenuItem[] | false
 export type OnError = (error: Error) => void
 export type OnFocus = () => void
 export type OnBlur = () => void
